[add] ユニット選択状態復元処理をユーティリティ関数に分離
コンポーネントの責務を分離し、コードの再利用性と可読性を向上させるため、選択されたユニットのユーザーと名前を復元する処理をsessionUtils.jsに移行しました。
